requirements.txt
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
httptools==0.1.1
minio==7.0.3
motor==2.4.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
pymongo==3.11.3
python-dotenv==0.17.0
python-jose==3.2.0
python-magic==0.4.22
python-multipart==0.0.5
PyYAML==5.4.1
rsa==4.7.2
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
